moved tech record methods to tech record service
